+++ test/typescript/selector-custom-types/i18nextT.test.ts
@@ -47,9 +47,7 @@ describe('i18next.t', () => {
         i18next.t(($) => $.inter, { val: 'asdf', ns: 'custom' }),
       ).toEqualTypeOf<'some {{val}}'>();
       expectTypeOf(i18next.t(($) => $.inter, { val: 'asdf' })).toEqualTypeOf<'some {{val}}'>();
-      expectTypeOf(
-        i18next.t(($) => $.qux, { val: 'asdf' }),
-      ).toEqualTypeOf<'some {{val, number}}'>();
+      expectTypeOf(i18next.t(($) => $.qux, { val: 123 })).toEqualTypeOf<'some {{val, number}}'>();
     });
 
     it('should work when using unescaped values', () => {